First Pass Functioning Polynomial Implementation
Generic univariate polynomial terms and complete polynomials. There are
no constraints on coefficient or exponent types. At the moment there is
an ugly overlap of Field and Ring implicits for the coefficient
algebra. In order to do basically anything for the polynomial,
including printing, there is a prerequisite that there is an implicit
order[E] in scope. Similarly there is a necessary show method to print
out the polynomials.